# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Volunteer API
  description: |
    REST API для платформы волонтёрских мероприятий в MAX.
    Все запросы (кроме `GET /api/health`) требуют валидный `initData` 
    из MAX WebApp для аутентификации пользователя.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Локальный сервер разработки
  - url: https://didntthinkofdomen.online
    description: Продакшн-сервер
...
paths:
...
  /api/auth/check:
    post:
      tags: [Auth]
      summary: Проверка текущего пользователя
      description: Возвращает данные пользователя MAX и его профиль, если он уже зарегистрирован.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitDataRequest'
      responses:
        '200':
          description: Результат проверки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      description: Создаёт пользователя и профиль (волонтёр или организация).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/InitDataRequest'
                - $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  role: { type: string, enum: [volunteer, organization_creator] }
                  maxBridgeId: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Пользователь уже зарегистрирован
  /api/auth/profile:
    patch:
      tags: [Auth]
      summary: Обновление профиля
      description: Обновляет редактируемые поля профиля в зависимости от роли.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/InitDataRequest'
                - $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: Профиль обновлён
          content:
            application/json:
              schema:
                type: object
                properties:
                  role: { type: string }
                  updated: { type: boolean }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/auth/verify-contact:
    post:
      tags: [Auth]
      summary: Сохранение номера телефона
      description: Сохраняет номер телефона волонтёра, полученный через `window.WebApp.requestContact()`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [initData, phone]
              properties:
                initData: { type: string }
                phone: { type: string, example: "+79991234567" }
      responses:
        '200':
          description: Телефон сохранён
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  phone: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/auth/upload/documents:
    post:
      tags: [Upload]
      summary: Загрузка документов организации
      description: ZIP-архив или PDF до 20 МБ.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                initData:
                  type: string
      responses:
        '200':
          description: Файл загружен
          content:
            application/json:
              schema:
                type: object
                properties:
                  filename: { type: string }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: Файл слишком большой
  /api/dadata/suggest:
    post:
      tags: [DaData]
      summary: Подсказки адресов
      description: |
        Возвращает до 6 вариантов адреса по введённой строке. 
        При сбое DaData возвращается `{ suggestions: [], degraded: true }`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [initData, query]
              properties:
                initData: { type: string }
                query: { type: string, example: "Москва Тверская" }
      responses:
        '200':
          description: Список подсказок
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestions:
                    type: array
                    items:
                      $ref: '#/components/schemas/AddressVariant'
                  degraded:
                    type: boolean
                    description: true если DaData недоступен
  /api/dadata/clean:
    post:
      tags: [DaData]
      summary: Очистка и нормализация адреса
      description: |
        Пытается нормализовать введённый адрес. 
        При успехе возвращает `{ address: AddressVariant, degraded: false }`.
        Если адрес не найден — `{ address: null, degraded: false }`.
        При сбое DaData — `{ address: null, degraded: true }`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [initData, address]
              properties:
                initData: { type: string }
                address: { type: string }
      responses:
        '200':
          description: Результат очистки
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    allOf:
                      - $ref: '#/components/schemas/AddressVariant'
                    nullable: true
                  degraded:
                    type: boolean
...
components:
  schemas:
    InitDataRequest:
      type: object
      required: [initData]
      properties:
        initData:
          type: string
          description: Подписанная строка initData из MAX WebApp
    MaxUser:
      type: object
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        username: { type: string, nullable: true }
        photoUrl: { type: string, nullable: true }
    VolunteerProfile:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        middleName: { type: string, nullable: true }
        birthDate: { type: string, format: date-time }
        city: { type: string, nullable: true }
        address: { type: string, nullable: true }
    OrganizationProfile:
      type: object
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        contacts: { type: string, nullable: true }
        socialMediaLink: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        isOfficial: { type: boolean }
        hasDocuments: { type: boolean }
        city: { type: string, nullable: true }
        address: { type: string, nullable: true }
    CheckResponse:
      type: object
      required: [isNewUser]
      properties:
        isNewUser:
          type: boolean
        maxUser:
          $ref: '#/components/schemas/MaxUser'
        user:
          type: object
          nullable: true
          properties:
            id: { type: integer }
            role: { type: string, enum: [volunteer, organization_creator] }
            maxBridgeId: { type: string }
            volunteer:
              $ref: '#/components/schemas/VolunteerProfile'
              nullable: true
            organization:
              $ref: '#/components/schemas/OrganizationProfile'
              nullable: true
    RegisterDto:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [volunteer, organization_creator]
        firstName: { type: string }
        lastName: { type: string }
        middleName: { type: string }
        birthDate: { type: string, format: date }
        city: { type: string }
        address: { type: string }
        phone: { type: string }
        organizationName: { type: string }
        description: { type: string }
        contacts: { type: string }
        socialMediaLink: { type: string }
        logoUrl: { type: string }
        documentsZipFilename: { type: string }
    UpdateProfileDto:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        middleName: { type: string }
        birthDate: { type: string, format: date }
        organizationName: { type: string }
        description: { type: string }
        contacts: { type: string }
        socialMediaLink: { type: string }
        logoUrl: { type: string }
        city: { type: string }
        address: { type: string }
    AddressVariant:
      type: object
      properties:
        value:
          type: string
          example: "г Москва, ул Тверская, д 1"
        city:
          type: string
          nullable: true
          example: "Москва"
...
    Error:
      type: object
      properties:
        message:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        statusCode: { type: integer }
  responses:
    BadRequest:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Невалидный initData
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'